+++ hdrezka/__init__.py
@@ -2,6 +2,7 @@ from __future__ import annotations
 
 from .api.ajax import Ajax
 from .api.search import Search
+from .post.page import Page
 from .stream import Player
 
-__all__ = ('Ajax', 'Player', 'Search')
+__all__ = ('Player', 'Search', 'Page', 'Ajax')